Refactored orderCalculatorTest and made minor modifications according to review so that it aligns to junit testing conventions
